create brainstorming directory in handlers and move some files to the directory
