+++ handlers/brainstorming/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+
+if __name__ == '__main__':
+    pass
